Improve RPM to automatically set up an SSH user for remote log collection
When the RPM is installed, it will now create an ‘olpcjournalprocessor’ user
and generate it an SSH key. It will enable that SSH key, but limit it to
executing the olpc-journal-processor script using SSH forced commands.

If the person installing the RPM takes a copy of the generated private key,
they can remotely access the school server to process the logs in future.

Note that this means we need to be careful to not introduce escalation
vulnerabilities (such as shell escapes) in olpc-journal-processor, or it
would be possible for someone to get shell access through the
olpcjournalprocessor user.